Remove dependency on botocore requests
Replaced botocore requests with standalone package due to:
https://aws.amazon.com/blogs/developer/removing-the-vendored-version-of-requests-from-botocore/
