+++ Readme.md
@@ -21,6 +21,7 @@ Enabling all features, enabling cloudtrail (not counting s3 log storage) and usi
 
 - Athena Alerter uses cloudtrail, which requires "All features" to be enabled on your AWS account. More on that can be found in AWS documentation - [Enabling all features](https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_org_support-all-features.html),
 - You should create a s3 bucket to store the lambda code,
+- Host your own Lambda Layer with python3's `requests` or e.g. use communities' from your region: https://github.com/keithrozario/Klayers/tree/master/deployments/python3.7/arns
 - Your user account needs to have permissions to create all the resources used by athena alerter. If you're missing some, stack creation will fail and you'll see which permission you're lacking in the cloudformation console.
 
 ## Deployment
@@ -36,6 +37,7 @@ To deploy Athena Alerter:
     - DYNAMODB_TABLE_NAME - name of dynamodb table which will be used for storing query data. The table should not exist and will be created for you. The name needs to match the QUERIES_TABLE parameter from settings.py,
     - SQS_QUEUE_NAME - name of sqs queue used for query finished events. The queue should not exists and will be created for you. The name needs to match the SQS_QUEUE_URL parameter from settings.py.
     - SNS_ANOMALY_DETECTION_TOPIC_NAME - name to which cloudwatch pushes events about anomaly detection in query sizes. This value is used to connect the topic as input to notifier lambda function (setting metrics is done separately in terraform) - see anomaly detection section below
+    - REQUESTS_LAYER - ARN address to Lambda Layer containing `requests`
     
 Note that S3 bucket names need to be globally unique (that means for all aws accounts).
