+++ requirements.txt
@@ -1,3 +1,3 @@
-jinja2==2.10
-boto3==1.9.79
-requests2==2.16.0
+jinja2==2.10.*
+boto3==1.12.*
+requests==2.22.*
